+++ crates/axiom-bench/benches/domain_bench.rs
@@ -98,11 +98,11 @@ fn make_arbiter(reflex_t: u8, assoc_t: u8, trace_count: usize) -> Arbiter {
     ];
     let mut domain_map = HashMap::new();
     for c in &configs {
-        domain_map.insert(c.domain_id as u32, *c);
+        domain_map.insert(c.domain_id, *c);
     }
     let mut arbiter = Arbiter::new(domain_map, COM::new());
     for (role, c) in configs.iter().enumerate() {
-        let _ = arbiter.register_domain(role as u8, c.domain_id as u32);
+        let _ = arbiter.register_domain(role as u8, c.domain_id);
     }
     arbiter.experience_mut().set_thresholds(reflex_t, assoc_t);
     for i in 0..trace_count {
